+++ examples/string.go
@@ -3,12 +3,14 @@ package main
 import (
 	"fmt"
 
-	prompt "github.com/segmentio/go-prompt"
+	"github.com/riteshpradhan/go-prompt"
 )
 
 func main() {
 	println("need your name!")
 	first := prompt.String("first")
 	last := prompt.String("last")
-	fmt.Printf("\nHello %s %s\n", first, last)
+	def := prompt.StringDefault("Want default", "defaultValue")
+	fmt.Printf("\nHello %s %s \n", first, last)
+	fmt.Println("Default value is : ", def)
 }